Added structs for input data and time structure.
Added struct 'InputData', which is used to store the defined data which the model is built on. A set format for the input data makes it easier to assure the validity and consistency of the data.

Also added the struct 'Temporals', which stores the timesteps in the model, as well as information on the length between timesteps.