Added residual costs to states
Added functionality to take into account the value of the states (storages) at the end of the optimization. The value per unit can be defined in the input data. If the defined value is positive, the contents of the state is calculated as a "gain", if negative, as a "loss".