Added checks for input data contents
Added fields to the input data indicating whether structures related to online functionality, reserves, piece wise efficiency, etc have to be created. It is unnecessary to create these structures if they are not needed, and they can also lead to unpredictable behaviour. These flags now need to be used in the rest of the code.