Update import_input_data.jl
Made improvements to the import input data code. Fixed types of parameters, as well as divided the code into a function reading the input data excel, and another building the input data based on the read data. Also made changes aimed at making it more convenient to build a model for only a part of the timeseries in the input data file.