Name gen_constraints.
The non-setpoint constraints are called gen_con, which is assigned
to mc["gen_constraints"].  It is a JuMP sparse array
indexed by the gc name, scenario and time.  This is different from the
previous mc["gen_constraints"], which was a Dict of Dicts;
the old one wan indexed with [c][(s, t)] and the new one with [c, s, t].
I could recreate the old structure but it did not seem to be used,
so I won't bother.