Rewrite some tuple generators with comprehensions.
I think this makes them more comprehensible.  Some of the results are now
generators instead of vectors, but for others there is code that requires
vectors.  That is usually filter.  I find it odd that filter does not accept
generators.  However, most uses of filter and map would be better written
with comprehensions instead.