Rename the plural validate_tuple to validate_tuples.
It was a bad overloading, distinguishing between a string tuple and
a vector of anything.  The plural version should really be made to work
with any kind of iterable, but tuples are also iterable.

This made me realise the sorry state of Julia dev tools.  VS Code should
be the best IDE (for Julia) but made an absolute mess of this
de-overloading, requiring a lot of manual fixing.  It is possible
that erroneous calls remain somewhere; there is no static analysis
to find them.  The tests pass but may not cover everything.