Switch to DateTime.
Timestamps are now DateTimes instead of strings.  The big exception is all
the tuples generated in tuples.jl, which continue to use strings because
tuples with DateTimes look ugly when converted to string, which happens
when they index JuMP objects.  In particular validate_tuple[s] still work
with strings.  For compatibility, Temporals.t remain strings.
Temporals.times is an OrderedDict with string keys and DateTime values,
providing easy access to both.

TimeZones has been removed from the project to ensure that we are no
longer using it.  If some use for it is later found, just add back.
For now, our only source of input is Excel workbooks, which do not
provide time zone information.

It is a mess because we could not completely transition to DateTime,
so now we have both them and strings.  We should move towards tuple-free
indexing of JuMP objects, particularly when time factors out from the
index set.  DateTimes work well as indices then.