Added flexible inflows
Added the "flex inflow" functionality to Predicer. The functionality allows the user to define time periods, during which the model can allocate a given amount of inflow (either positive or negative) to a node.

The changes include adding a new field (flex_inflow) to the Node struct, a function for building the index tuple, function to build variables, relevant constraints, as well as importing the input data from the input data file. Added some rudimentary data validation for flex inflows as well.